Finalize helper methods, update test project to use the preview helpers
