+++ FirebaseInAppMessaging/Swift/Source/SwiftUIPreviewHelpers.swift
@@ -7,24 +7,25 @@
 
 import FirebaseInAppMessaging
 
-struct InAppMessagingPreviewHelpers {
-  static func cardMessage(campaignName: String = "Card message campaign",
-                          title: String = "Title for modal message",
-                          body: String? = "Body for modal message",
-                          textColor: UIColor = UIColor.label,
-                          backgroundColor: UIColor = UIColor.systemBackground,
-                          portraitImage: UIImage = UIImage(systemName: "message")!,
-                          landscapeImage: UIImage? = UIImage(systemName: "message.fill"),
-                          primaryButtonText: String = "Click me!",
-                          primaryButtonTextColor: UIColor = UIColor.systemBlue,
-                          primaryButtonBackgroundColor: UIColor = UIColor.secondarySystemBackground,
-                          primaryActionURL: URL? = nil,
-                          secondaryButtonText: String? = "Dismiss",
-                          secondaryButtonTextColor: UIColor? = UIColor.secondaryLabel,
-                          secondaryButtonBackgroundColor: UIColor? = UIColor
-                            .tertiarySystemBackground,
-                          secondaryActionURL: URL? = nil,
-                          appData: [String: String]? = nil) -> InAppMessagingCardDisplay {
+public struct InAppMessagingPreviewHelpers {
+  public static func cardMessage(campaignName: String = "Card message campaign",
+                                 title: String = "Title for modal message",
+                                 body: String? = "Body for modal message",
+                                 textColor: UIColor = UIColor.label,
+                                 backgroundColor: UIColor = UIColor.systemBackground,
+                                 portraitImage: UIImage = UIImage(systemName: "message")!,
+                                 landscapeImage: UIImage? = UIImage(systemName: "message.fill"),
+                                 primaryButtonText: String = "Click me!",
+                                 primaryButtonTextColor: UIColor = UIColor.systemBlue,
+                                 primaryButtonBackgroundColor: UIColor = UIColor
+                                   .secondarySystemBackground,
+                                 primaryActionURL: URL? = nil,
+                                 secondaryButtonText: String? = "Dismiss",
+                                 secondaryButtonTextColor: UIColor? = UIColor.secondaryLabel,
+                                 secondaryButtonBackgroundColor: UIColor? = UIColor
+                                   .tertiarySystemBackground,
+                                 secondaryActionURL: URL? = nil,
+                                 appData: [String: String]? = nil) -> InAppMessagingCardDisplay {
     // This may crash the preview if an invalid portrait image is provided, card messages must have
     // a valid portrait image.
     let portraitImageData = InAppMessagingImageData(imageURL: "http://fakeurl.com",
@@ -65,18 +66,18 @@ struct InAppMessagingPreviewHelpers {
     )
   }
 
-  static func modalMessage(campaignName: String = "Modal message campaign",
-                           title: String = "Title for modal message",
-                           body: String? = "Body for modal message",
-                           textColor: UIColor = UIColor.black,
-                           backgroundColor: UIColor = UIColor.white,
-                           image: UIImage? = UIImage(systemName: "message"),
-                           buttonText: String? = "Click me!",
-                           buttonTextColor: UIColor? = UIColor.systemBlue,
-                           buttonBackgroundColor: UIColor? = UIColor
-                             .white,
-                           actionURL: URL? = nil,
-                           appData: [String: String]? = nil) -> InAppMessagingModalDisplay {
+  public static func modalMessage(campaignName: String = "Modal message campaign",
+                                  title: String = "Title for modal message",
+                                  body: String? = "Body for modal message",
+                                  textColor: UIColor = UIColor.black,
+                                  backgroundColor: UIColor = UIColor.white,
+                                  image: UIImage? = UIImage(systemName: "message"),
+                                  buttonText: String? = "Click me!",
+                                  buttonTextColor: UIColor? = UIColor.systemBlue,
+                                  buttonBackgroundColor: UIColor? = UIColor
+                                    .white,
+                                  actionURL: URL? = nil,
+                                  appData: [String: String]? = nil) -> InAppMessagingModalDisplay {
     var imageData: InAppMessagingImageData?
     if let data = image?.pngData() {
       imageData = InAppMessagingImageData(imageURL: "http://fakeurl.com", imageData: data)
@@ -103,14 +104,15 @@ struct InAppMessagingPreviewHelpers {
     )
   }
 
-  static func bannerMessage(campaignName: String = "Banner message campaign",
-                            title: String = "Title for banner message",
-                            body: String? = "Body for banner message",
-                            textColor: UIColor = UIColor.black,
-                            backgroundColor: UIColor = UIColor.white,
-                            image: UIImage? = UIImage(systemName: "message"),
-                            actionURL: URL? = nil,
-                            appData: [String: String]? = nil) -> InAppMessagingBannerDisplay {
+  public static func bannerMessage(campaignName: String = "Banner message campaign",
+                                   title: String = "Title for banner message",
+                                   body: String? = "Body for banner message",
+                                   textColor: UIColor = UIColor.black,
+                                   backgroundColor: UIColor = UIColor.white,
+                                   image: UIImage? = UIImage(systemName: "message"),
+                                   actionURL: URL? = nil,
+                                   appData: [String: String]? = nil)
+    -> InAppMessagingBannerDisplay {
     var imageData: InAppMessagingImageData?
     if let data = image?.pngData() {
       imageData = InAppMessagingImageData(imageURL: "http://fakeurl.com", imageData: data)
@@ -127,10 +129,10 @@ struct InAppMessagingPreviewHelpers {
     )
   }
 
-  static func imageOnlyMessage(campaignName: String = "Image-only message campaign",
-                               image: UIImage,
-                               actionURL: URL? = nil,
-                               appData: [String: String]? = nil)
+  public static func imageOnlyMessage(campaignName: String = "Image-only message campaign",
+                                      image: UIImage,
+                                      actionURL: URL? = nil,
+                                      appData: [String: String]? = nil)
     -> InAppMessagingImageOnlyDisplay {
     // This may crash the preview if an invalid image is provided, image-only messages must have a
     // valid portrait image.
@@ -143,4 +145,6 @@ struct InAppMessagingPreviewHelpers {
       appData: appData
     )
   }
+
+  public class Delegate: NSObject, InAppMessagingDisplayDelegate {}
 }
